move slide_limiter to library
