+++ ydb/core/tx/limiter/ya.make
@@ -1,5 +1,3 @@
 RECURSE(
     grouped_memory
-    service
-    usage
 )
