+++ ydb/library/slide_limiter/service/ya.make
@@ -5,7 +5,7 @@ SRCS(
 )
 
 PEERDIR(
-    ydb/core/tx/limiter/usage
+    ydb/library/slide_limiter/usage
     ydb/core/protos
 )
