+++ ydb/library/slide_limiter/ya.make
@@ -1,0 +1,4 @@
+RECURSE(
+    service
+    usage
+)
